% Generated by roxygen2 (4.0.2): do not edit by hand
\name{begin_end}
\alias{begin_end}
\title{Function to query ranges of MLG observations.}
\usage{
begin_end(x)
}
\arguments{
\item{x}{a logical vector of a single MLG observations per year.}
}
\value{
two integers indicating the indices for the first and last observed
instances of the MLG.
}
\description{
In terms of what I use it for: To find the first and last date a MLG was found.
}
\keyword{internal}

