% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_loc_table}
\alias{calc_loc_table}
\title{Calculate allelic or genotypic diversity by population.}
\usage{
calc_loc_table(dat, hier, combine = FALSE, lev = "allele", cc = TRUE)
}
\arguments{
\item{dat}{a genind or genclone object}

\item{hier}{a hierarchy to set for clone correction}

\item{combine}{logical. Should the hierarchy be combined to create the population after clone correction (default, FALSE, no).}

\item{lev}{see \code{\link[poppr]{locus_table}}.}

\item{cc}{logical. If TRUE (default), the data is clone corrected by the hierarchy
before stats are calculated.}
}
\value{
a list containing \itemize{
\item{pops - }{a list of locus tables by population}
\item{total - }{a locus table for the pooled population}
}
}
\description{
Calculate allelic or genotypic diversity by population.
}
\examples{
library(poppr)
data(Pinf)
calc_loc_table(Pinf, hier = ~Continent/Country)
}
\author{
Zhian N. Kamvar
}

