% Generated by roxygen2 (4.0.2): do not edit by hand
\name{char2pal}
\alias{char2pal}
\title{Creates a named color palette.}
\usage{
char2pal(x, pal = adegenet::funky)
}
\arguments{
\item{x}{a vector of identifiers to be used for colors.}

\item{pal}{a color palette. Default is adegenet's "funky"}
}
\value{
a named character vector of hexadecimal colors.
}
\description{
This is useful for defining a color palette that can be used by population factors.
}
\examples{
char2pal(LETTERS)
}

