% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{comparePal}
\alias{comparePal}
\title{Color palette for 9 populations defined in forest and nursery isolates of
\emph{Phytophthora ramorum}.}
\format{a named character vector of hexadecimal colors derived from the
  function \code{\link{char2pal}} with the \pkg{adegenet} palette
  \code{\link[adegenet]{funky}}.}
\usage{
data(comparePal)
}
\description{
This palette defines a separate color for each popualtion in the
  data set \code{\link{for2nur}}. This is based on the hierarchy of
  \code{~SOURCE/STATE}. Forest populations are defined by colors, whereas
  Nursery populations are defined by shades of grey/black. This is especially
  useful for the function \code{\link[ggplot2]{scale_color_manual}} in the
  \pkg{ggplot2} package.
}
\references{
ZN Kamvar, MM Larsen, AM Kanaskie, EM Hansen  and NJ Grünwald.
  20XX. Spatial and temporal population dynamics of the sudden oak death
  epidemic in Oregon Forests. Phytopathology XX:XXX-XXX.
}

