% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{for2nur}
\alias{for2nur}
\title{\emph{Phytophthora ramorum} forest and nursery genotype data.}
\format{a \code{\link{genclone}} object.}
\usage{
data(for2nur)
}
\description{
This data set contains 729 diploid samples from the forest in
  Curry County, OR, USA between 2001 and 2014 and genotypes found in
  Nurseries from CA and OR between 2000 and 2012. A population hierarchy
  contains the names SOURCE, YEAR, and STATE. SOURCE represents either
  watershed regions (forest data) or nursery. YEAR represents the year
  sampled. STATE represents the state of origin (OR or CA). The \code{other}
  slot contains xy coordinates for each sample ($xy), the population
  hierarchy in a data frame ($population_hierarchy), and the repeat motif
  lengths for each of the 5 loci ($REPLEN).
}
\references{
ZN Kamvar, MM Larsen, AM Kanaskie, EM Hansen  and NJ Grünwald.
  20XX. Spatial and temporal population dynamics of the sudden oak death
  epidemic in Oregon Forests. Phytopathology XX:XXX-XXX.
}

