% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_layout}
\alias{get_layout}
\title{Generate layout function}
\usage{
get_layout(graph, LAYOUT = layout.auto)
}
\arguments{
\item{graph}{an igraph object}

\item{LAYOUT}{a layout function defined by igraph (default) OR a matrix
defining the layout.}
}
\value{
a function that will return the a layout matrix based on the layout
function.
}
\description{
Generate layout function
}
\details{
This function is used for generating a function that will return a
pre-defined layout. This is useful for subsetting networks but wanting all of
the points to be equivalent. Used with the layfun argument of plot_poppr_msn
}

