% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_ranges}
\alias{get_ranges}
\title{Get bounding box from Latitute and Lognitude vectors.}
\usage{
get_ranges(LAT, LON)
}
\arguments{
\item{LAT}{a numeric vector of latitudes}

\item{LON}{a numeric vector of longitudes}
}
\value{
a numeric vector of four elements containing:
\enumerate{
\item lower left LON
\item lower left LAT
\item upper right LON
\item upper right LAT}
}
\description{
Get bounding box from Latitute and Lognitude vectors.
}

