% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lazierLoad}
\alias{lazierLoad}
\title{Performs lazy load on a directory}
\usage{
lazierLoad(path)
}
\arguments{
\item{path}{a filepath containing the necessary files for lazy loading}
}
\description{
Performs lazy load on a directory
}
\details{
This function will go into a directory, search for all the files
that seem like they can be lazily loaded and attempt to load them.
}
\examples{
\dontrun{
lazierLoad("Diversity_stats_cache/html")
}
}

