% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match_duplicates}
\alias{match_duplicates}
\title{match and return all duplicated elements.}
\usage{
match_duplicates(x)
}
\arguments{
\item{x}{a character or numeric vector}
}
\value{
a logical vector of elements that are duplicated.
}
\description{
match and return all duplicated elements.
}

