% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge_duplicates}
\alias{merge_duplicates}
\title{Merge or remove duplicate entries from a genclone object}
\usage{
merge_duplicates(x, stats = TRUE)
}
\arguments{
\item{x}{a genind object}

\item{stats}{if set to \code{TRUE}, the status of the merges will be printed
to the console.}
}
\value{
a genind object
}
\description{
This will compare duplicate names in genclone objects and merge the ones with the
same genotype and remove the ones with differing genotypes
}

