% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mlg.barplot}
\alias{mlg.barplot}
\title{Plot MLG distribution by year.}
\usage{
mlg.barplot(mat, total = TRUE, pal = funky)
}
\arguments{
\item{mat}{a matrix derived from \code{\link[poppr]{mlg.table}}}

\item{total}{defaults to \code{TRUE} if there is a row for total MLG counts}

\item{pal}{a color palette or vector of named colors to be used for MLGs.}
}
\value{
a ggplot object that contains a barplot of MLG abundances per year.
}
\description{
Plot MLG distribution by year.
}

