% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mlg.heatmap}
\alias{mlg.heatmap}
\title{Plot MLG occurance by population (year)}
\usage{
mlg.heatmap(mat, total = TRUE, log_transform = TRUE)
}
\arguments{
\item{mat}{a matrix derived from \code{\link[poppr]{mlg.table}}}

\item{total}{defaults to \code{TRUE} if there is a row for total MLG counts}

\item{log_transform}{should the counts be log transformed for easier viewing? Default to TRUE.}
}
\value{
a ggplot object that contains a heatmap of multilocus genotypes
per year.
}
\description{
Plot MLG occurance by population (year)
}

