% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MLG_range}
\alias{MLG_range}
\title{Transform MLG matrix into date range of observations per MLG}
\usage{
MLG_range(mat)
}
\arguments{
\item{mat}{a TRUE/FALSE matrix of MLGs (columns) per year (rows)}
}
\value{
a character matrix of two columns and rows equal to the
number of observed MLGs giving the first and last observed instances.
}
\description{
Transform MLG matrix into date range of observations per MLG
}

