% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mlgFromString}
\alias{mlgFromString}
\title{Function to return numeric MLGs from mlg strings.}
\usage{
mlgFromString(x)
}
\arguments{
\item{x}{a string of MLG definitions formatted as "MLG.n"
where n is a number.}
}
\value{
a numeric vector defining the multilocus genotypes.
}
\description{
Yup
}
\examples{
mlgFromString("MLG.100")
}
\author{
Zhian N. Kamvar
}

