% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{myPal}
\alias{myPal}
\title{Color palette for 70 Multilocus Genotypes in \emph{Phytophthora ramorum}
forest data.}
\format{a named character vector of hexadecimal colors derived from the
  function \code{\link{char2pal}} with the \pkg{adegenet} palette
  \code{\link[adegenet]{funky}}.}
\usage{
data(myPal)
}
\description{
This palette defines a separate color for each MLG in the data
  set \code{\link{ramdat}}. This is especially useful for the function
  \code{\link[ggplot2]{scale_color_manual}} in the \pkg{ggplot2} package.
}
\references{
ZN Kamvar, MM Larsen, AM Kanaskie, EM Hansen  and NJ Grünwald.
  20XX. Spatial and temporal population dynamics of the sudden oak death
  epidemic in Oregon Forests. Phytopathology XX:XXX-XXX.
}

