% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pad_zeroes}
\alias{pad_zeroes}
\title{Add leading zeroes onto a number range.}
\usage{
pad_zeroes(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a character vector where all numbers have the same
number of characters.
}
\description{
Add leading zeroes onto a number range.
}

