% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_loc_table}
\alias{plot_loc_table}
\title{Visualize locus tables}
\usage{
plot_loc_table(tab)
}
\arguments{
\item{tab}{a list of locus tables produced from \code{\link{calc_loc_table}}.}
}
\value{
a ggplot
}
\description{
Visualize locus tables
}
\examples{
library(poppr)
data(Pinf)
plot_loc_table(calc_loc_table(Pinf, hier = ~Continent/Country))
}
\author{
Zhian N. Kamvar
}

