% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior values from DAPC analysis in adegenet}
\usage{
plot_posterior(da.object, gid, pal = funky, cols = 1)
}
\arguments{
\item{da.object}{an object of class "dapc"}

\item{gid}{an object of class "genind"}

\item{pal}{a color palette}

\item{cols}{the number of columns to display}
}
\value{
a ggplot object with each population stacked on top of each other.
}
\description{
Plot posterior values from DAPC analysis in adegenet
}

