% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{Metadata for the forest populations of \emph{Phytophthora ramorum}.}
\format{a \code{\link{data.frame}} with 9 columns.}
\usage{
data(pop_data)
}
\description{
This data frame contains 9 columns containing metadata that
  pretains to the data contained in \code{\link{ramdat}}: \itemize{
  \item{ISOLATE - }{Original names given to each isolate. Decimal notations
  represent isolates collected from same host.} \item{SAMPLE - }{Indicator of
  host sample each isolate originated from} \item{LAT - }{Lattitude
  measurements for each isolate} \item{LON - }{Longitude measurements for
  each isolate} \item{NOTES - }{Notes for specific isolates} \item{ZONE1 -
  }{Watershed classifications at a fine scale} \item{ZONE2 - }{Watershed
  classifications at a broad scale} \item{YEAR - }{Year isolated} \item{MLG -
  }{Multilocus genotype assignment as defined in poppr}}
}
\references{
ZN Kamvar, MM Larsen, AM Kanaskie, EM Hansen  and NJ Grünwald.
  20XX. Spatial and temporal population dynamics of the sudden oak death
  epidemic in Oregon Forests. Phytopathology XX:XXX-XXX.
}

