% Generated by roxygen2 (4.0.2): do not edit by hand
\name{population_subgraph}
\alias{population_subgraph}
\title{Highlight a subpopulation in a minimum spaning network}
\usage{
population_subgraph(x, poppr_msn, ..., sublist = "ALL", marker = TRUE)
}
\arguments{
\item{x}{a genind or genclone object.}

\item{poppr_msn}{a minimum spanning network derived from \code{x}.}

\item{...}{arguments to be passed on to \code{plot_poppr_msn}}

\item{sublist}{a character string indicating the subpopulation to plot.}

\item{marker}{should the samples within the subgraph be marked with a polygon
  underneath the nodes?}
}
\description{
Highlight a subpopulation in a minimum spaning network
}

