% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepare_mlg_df}
\alias{prepare_mlg_df}
\title{Rearrange mlg.table output into a data frame for plotting.}
\usage{
prepare_mlg_df(mat, total = TRUE, type = c("heatmap", "identity"))
}
\arguments{
\item{mat}{a matrix derived from \code{\link[poppr]{mlg.table}}}

\item{total}{defaults to \code{TRUE} if there is a row for total MLG counts}

\item{type}{"heatmap" (default) will set all zero values to NA. "identity"
will keep zeroes as zero.}
}
\value{
a molten data frame with the values of "MLG", "count", and "unique".
}
\description{
Rearrange mlg.table output into a data frame for plotting.
}
\keyword{internal}

