% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ramdat}
\alias{ramdat}
\title{\emph{Phytophthora ramorum} forest genotype data.}
\format{a \code{\link{genclone}} object.}
\usage{
data(ramdat)
}
\description{
This data set contains 513 diploid samples from the forest in
  Curry County, OR, USA between 2001 and 2014. A population hierarchy
  contains the names Pop, ZONE1, and ZONE2. Pop represents years, ZONE2
  represents broad-scale watershed regions, and ZONE1 represents finer-scale
  watershed regions. The \code{other} slot contains xy coordinates for each
  sample ($xy), the population hierarchy in a data frame
  ($population_hierarchy), and the repeat motif lengths for each of the 5
  loci ($REPLEN).
}
\references{
ZN Kamvar, MM Larsen, AM Kanaskie, EM Hansen  and NJ Grünwald.
  20XX. Spatial and temporal population dynamics of the sudden oak death
  epidemic in Oregon Forests. Phytopathology XX:XXX-XXX.
}

