% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_the_edges}
\alias{remove_the_edges}
\title{Removes edges not shared in both graphs}
\usage{
remove_the_edges(a, b)
}
\arguments{
\item{a}{an igraph graph object}

\item{b}{an igraph graph object with a subset of nodes from a OR a character
labeling the nodes to subset.}
}
\value{
an modified version of graph \code{a} only containing edges from
  graph \code{b}
}
\description{
This function will take two graphs, \code{a} and \code{b}, where \code{b} is
a subgraph of \code{a} and remove all edges in \code{a} that are not present
in \code{b}. Note that there can be edges in \code{b} that are not present in
\code{a}. This function will only return edges present in \code{a}.
}
\keyword{internal}

