% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resetMLG}
\alias{resetMLG}
\title{Function to reset MLG counts.}
\usage{
resetMLG(x)
}
\arguments{
\item{x}{genclone object.}
}
\value{
a genclone object with reset MLG counts.
}
\description{
This is a workaround for attempting to subset populations before running
AMOVA on genclone objects.
}
\examples{
data(Pinf)
resetMLG(Pinf)
}
\author{
Zhian N. Kamvar
}

