% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spatial_stats}
\alias{spatial_stats}
\title{Spatial test on multiple populations.}
\usage{
spatial_stats(gen, xy, stat = "mantel", hierarchy = NULL,
  distance = "nei.dist", sample = 999, plot = TRUE, ncol = 3,
  seed = NULL, ...)
}
\arguments{
\item{gen}{a genclone object}

\item{xy}{a matrix of latitude and longitude data or a distance matrix.}

\item{stat}{a character selectine either "mantel" or "lm"}

\item{hierarchy}{a formula specifying the hierarchical levels to be used to
define the population.}

\item{distance}{a function to caclulate the genetic distance or a genetic
distance matrix of class dist.}

\item{sample}{the number of replications for each mantel test (performed in
C)}

\item{plot}{if TRUE, all possible}

\item{ncol}{if plot = TRUE, how many columns should be plotted?}

\item{seed}{numeric This is he seed to set before randomizations for the
mantel test. By default, this is set to NULL so that the results are
random. When given a numeric value, results are reproducible.}

\item{...}{arguments passed onto the distance function.}
}
\value{
a list containing multiple test results.
}
\description{
Performs a mantel test or linear model on multiple populations in a genclone
object.
}

