% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_replen}
\alias{test_replen}
\title{Test repeat length consistency.}
\usage{
test_replen(gid, replen)
}
\arguments{
\item{gid}{a genind object}

\item{replen}{a numeric vector of repeat motif lengths.}
}
\value{
a logical vector indicating whether or not the repeat motif length
is consistent.
}
\description{
This function will test for consistency in the sense that all alleles are
able to be represented as discrete units after division and rounding.
}
\examples{
library(poppr)
data(nancycats)
test_replen(nancycats, rep(2, 9))
}

