% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update_my_graph}
\alias{update_my_graph}
\title{Function used to update colors in poppr msn}
\usage{
update_my_graph(graphlist, PALETTE)
}
\arguments{
\item{graphlist}{a graph produced from one of poppr's *.msn functions}

\item{PALETTE}{a function or character defining a palette}
}
\value{
a shiny new graph
}
\description{
Function used to update colors in poppr msn
}

